feat: move tests in a test folder
